Updated Cocos2D to 3.0RC4 (fixed the bug where it would crash on 64bit platform in XCode 5.1)

+++ Strik.spritebuilder/STKAvatarPickerController.m
@@ -107,7 +107,7 @@
 {
 	// Determine range for page
 	int start = page * [STKAvatarPage avatarsPerPage];
-	int length = MIN(self.allAvatars.count - start, [STKAvatarPage avatarsPerPage]);
+	int length = (int)MIN(self.allAvatars.count - start, [STKAvatarPage avatarsPerPage]);
 	
 	NSRange avatarRange = NSMakeRange(start, length);
